+++ check-npm.js
@@ -3,5 +3,4 @@
 // checkNpmVersions({
 //   "react": ">=0.14.7 || ^15.0.0-rc.2",
 //   "react-dom": ">=0.14.7 || ^15.0.0-rc.2",
-//   "tracker-component": "^1.3.13"
 // });