+++ imports/helpers.js
@@ -18,7 +18,7 @@ export function getLoginServices() {
   // backwards-compatibility.
   services.sort();
 
-  return _.map(services, function(name){
+  return services.map(function(name){
     return {name: name};
   });
 };
@@ -64,10 +64,10 @@ export function validateEmail(email, showMessage, clearMessage) {
   if (Accounts.ui._options.emailPattern.test(email)) {
     return true;
   } else if (!email || email.length === 0) {
-    showMessage(T9n.get("error.emailRequired"), 'warning', false, 'email');
+    showMessage("error.emailRequired", 'warning', false, 'email');
     return false;
   } else {
-    showMessage(T9n.get("error.accounts.Invalid email"), 'warning', false, 'email');
+    showMessage("error.accounts.Invalid email", 'warning', false, 'email');
     return false;
   }
 }
@@ -76,7 +76,8 @@ export function validatePassword(password = '', showMessage, clearMessage){
   if (password.length >= Accounts.ui._options.minimumPasswordLength) {
     return true;
   } else {
-    const errMsg = T9n.get("error.minChar").replace(/7/, Accounts.ui._options.minimumPasswordLength);
+    // const errMsg = T9n.get("error.minChar").replace(/7/, Accounts.ui._options.minimumPasswordLength);
+    const errMsg = "error.minChar"
     showMessage(errMsg, 'warning', false, 'password');
     return false;
   }
@@ -87,7 +88,7 @@ export function validateUsername(username, showMessage, clearMessage, formState)
     return true;
   } else {
     const fieldName = (passwordSignupFields() === 'USERNAME_ONLY' || formState === STATES.SIGN_UP) ? 'username' : 'usernameOrEmail';
-    showMessage(T9n.get("error.usernameRequired"), 'warning', false, fieldName);
+    showMessage("error.usernameRequired", 'warning', false, fieldName);
     return false;
   }
 }