+++ imports/ui/components/Button.jsx
@@ -1,15 +1,11 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Accounts } from 'meteor/accounts-base';
-import PropTypes from 'prop-types'
 
 let Link;
 try { Link = require('react-router').Link; } catch(e) {}
 
 export class Button extends React.Component {
-  propTypes: {
-    onClick: PropTypes.func
-  }
-
   render () {
     const {
       label,
@@ -33,5 +29,9 @@ export class Button extends React.Component {
                    onClick={ onClick }>{ label }</button>;
   }
 }
+
+Button.propTypes = {
+  onClick: PropTypes.func
+};
 
 Accounts.ui.Button = Button;