+++ imports/ui/components/Form.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import { Accounts } from 'meteor/accounts-base';
-import PropTypes from 'prop-types'
 
 import './Fields.jsx';
 import './Buttons.jsx';
@@ -11,14 +11,6 @@ import './SocialButtons.jsx';
 import './FormMessages.jsx';
 
 export class Form extends React.Component {
-  propTypes: {
-    oauthServices: PropTypes.object,
-    fields: PropTypes.object.isRequired,
-    buttons: PropTypes.object.isRequired,
-    error: PropTypes.string,
-    ready: PropTypes.bool
-  };
-
   componentDidMount() {
     let form = this.form;
     if (form) {
@@ -36,6 +28,7 @@ export class Form extends React.Component {
       buttons,
       error,
       messages,
+      translate,
       ready = true,
       className
     } = this.props;
@@ -48,12 +41,21 @@ export class Form extends React.Component {
       >
         <Accounts.ui.Fields fields={ fields } />
         <Accounts.ui.Buttons buttons={ buttons } />
-        <Accounts.ui.PasswordOrService oauthServices={ oauthServices } />
+        <Accounts.ui.PasswordOrService oauthServices={ oauthServices } translate={ translate } />
         <Accounts.ui.SocialButtons oauthServices={ oauthServices } />
         <Accounts.ui.FormMessages messages={messages} />
       </form>
     );
   }
 }
+
+Form.propTypes = {
+  oauthServices: PropTypes.object,
+  fields: PropTypes.object.isRequired,
+  buttons: PropTypes.object.isRequired,
+  translate: PropTypes.func.isRequired,
+  error: PropTypes.string,
+  ready: PropTypes.bool
+};
 
 Accounts.ui.Form = Form;