+++ imports/ui/components/LoginForm.jsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { PropTypes, Component } from 'react';
 import ReactDOM from 'react-dom';
-import Tracker from 'tracker-component';
+import { createContainer } from 'meteor/react-meteor-data';
 import { Accounts } from 'meteor/accounts-base';
 import { T9n } from 'meteor/softwarerero:accounts-t9n';
 import { KEY_PREFIX } from '../../login_session.js';
@@ -18,7 +18,15 @@ import {
   capitalize
 } from '../../helpers.js';
 
-export class LoginForm extends Tracker.Component {
+function indexBy(array, key) {
+  const result = {};
+  array.forEach(function(obj) {
+    result[obj[key]] = obj;
+  });
+  return result;
+}
+
+class LoginForm extends Component {
   constructor(props) {
     super(props);
     let {
@@ -45,6 +53,7 @@ export class LoginForm extends Tracker.Component {
       onPreSignUpHook: props.onPreSignUpHook || Accounts.ui._options.onPreSignUpHook,
       onPostSignUpHook: props.onPostSignUpHook || Accounts.ui._options.onPostSignUpHook,
     };
+    this.translate = this.translate.bind(this);
   }
 
   componentDidMount() {
@@ -71,22 +80,11 @@ export class LoginForm extends Tracker.Component {
         Session.set(KEY_PREFIX + 'state', null);
         break;
     }
-    
+
     // Add default field values once the form did mount on the client
     this.setState(prevState => ({
       ...this.getDefaultFieldValues(),
     }));
-    
-    // Listen for the user to login/logout.
-    this.autorun(() => {
-      
-      // Add the services list to the user.
-      this.subscribe('servicesList');
-      this.setState({
-        user: Accounts.user()
-      });
-      
-    });
   }
 
   componentWillReceiveProps(nextProps, nextContext) {
@@ -99,11 +97,18 @@ export class LoginForm extends Tracker.Component {
   }
 
   componentDidUpdate(prevProps, prevState) {
-    if (!prevState.user !== !this.state.user) {
+    if (!prevProps.user !== !this.props.user) {
       this.setState({
-        formState: this.state.user ? STATES.PROFILE : STATES.SIGN_IN
+        formState: this.props.user ? STATES.PROFILE : STATES.SIGN_IN
       });
     }
+  }
+
+  translate(text) {
+    // if (this.props.t) {
+    //   return this.props.t(text);
+    // }
+    return T9n.get(text);
   }
 
   validateField(field, value) {
@@ -131,8 +136,8 @@ export class LoginForm extends Tracker.Component {
   getUsernameOrEmailField() {
     return {
       id: 'usernameOrEmail',
-      hint: T9n.get('enterUsernameOrEmail'),
-      label: T9n.get('usernameOrEmail'),
+      hint: this.translate('enterUsernameOrEmail'),
+      label: this.translate('usernameOrEmail'),
       required: true,
       defaultValue: this.state.username || "",
       onChange: this.handleChange.bind(this, 'usernameOrEmail'),
@@ -143,8 +148,8 @@ export class LoginForm extends Tracker.Component {
   getUsernameField() {
     return {
       id: 'username',
-      hint: T9n.get('enterUsername'),
-      label: T9n.get('username'),
+      hint: this.translate('enterUsername'),
+      label: this.translate('username'),
       required: true,
       defaultValue: this.state.username || "",
       onChange: this.handleChange.bind(this, 'username'),
@@ -155,8 +160,8 @@ export class LoginForm extends Tracker.Component {
   getEmailField() {
     return {
       id: 'email',
-      hint: T9n.get('enterEmail'),
-      label: T9n.get('email'),
+      hint: this.translate('enterEmail'),
+      label: this.translate('email'),
       type: 'email',
       required: true,
       defaultValue: this.state.email || "",
@@ -168,8 +173,8 @@ export class LoginForm extends Tracker.Component {
   getPasswordField() {
     return {
       id: 'password',
-      hint: T9n.get('enterPassword'),
-      label: T9n.get('password'),
+      hint: this.translate('enterPassword'),
+      label: this.translate('password'),
       type: 'password',
       required: true,
       defaultValue: this.state.password || "",
@@ -181,8 +186,8 @@ export class LoginForm extends Tracker.Component {
   getSetPasswordField() {
     return {
       id: 'newPassword',
-      hint: T9n.get('enterPassword'),
-      label: T9n.get('choosePassword'),
+      hint: this.translate('enterPassword'),
+      label: this.translate('choosePassword'),
       type: 'password',
       required: true,
       onChange: this.handleChange.bind(this, 'newPassword')
@@ -192,8 +197,8 @@ export class LoginForm extends Tracker.Component {
   getNewPasswordField() {
     return {
       id: 'newPassword',
-      hint: T9n.get('enterNewPassword'),
-      label: T9n.get('newPassword'),
+      hint: this.translate('enterNewPassword'),
+      label: this.translate('newPassword'),
       type: 'password',
       required: true,
       onChange: this.handleChange.bind(this, 'newPassword'),
@@ -225,11 +230,11 @@ export class LoginForm extends Tracker.Component {
     }
 
     if (hasPasswordService() && formState == STATES.SIGN_IN) {
-      if (_.contains([
+      if ([
         "USERNAME_AND_EMAIL",
         "USERNAME_AND_OPTIONAL_EMAIL",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getUsernameOrEmailField());
       }
 
@@ -237,48 +242,48 @@ export class LoginForm extends Tracker.Component {
         loginFields.push(this.getUsernameField());
       }
 
-      if (_.contains([
+      if ([
         "EMAIL_ONLY",
         "EMAIL_ONLY_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getEmailField());
       }
 
-      if (!_.contains([
+      if (![
         "EMAIL_ONLY_NO_PASSWORD",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getPasswordField());
       }
     }
 
     if (hasPasswordService() && formState == STATES.SIGN_UP) {
-      if (_.contains([
+      if ([
         "USERNAME_AND_EMAIL",
         "USERNAME_AND_OPTIONAL_EMAIL",
         "USERNAME_ONLY",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getUsernameField());
       }
 
-      if (_.contains([
+      if ([
         "USERNAME_AND_EMAIL",
         "EMAIL_ONLY",
         "EMAIL_ONLY_NO_PASSWORD",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getEmailField());
       }
 
-      if (_.contains(["USERNAME_AND_OPTIONAL_EMAIL"], passwordSignupFields())) {
+      if (["USERNAME_AND_OPTIONAL_EMAIL"].includes(passwordSignupFields())) {
         loginFields.push(Object.assign(this.getEmailField(), {required: false}));
       }
 
-      if (!_.contains([
+      if (![
         "EMAIL_ONLY_NO_PASSWORD",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())) {
+      ].includes(passwordSignupFields())) {
         loginFields.push(this.getPasswordField());
       }
     }
@@ -297,8 +302,7 @@ export class LoginForm extends Tracker.Component {
     if (this.showEnrollAccountForm()) {
       loginFields.push(this.getSetPasswordField());
     }
-
-    return _.indexBy(loginFields, 'id');
+    return indexBy(loginFields, 'id');
   }
 
   buttons() {
@@ -309,13 +313,14 @@ export class LoginForm extends Tracker.Component {
       changePasswordPath = Accounts.ui._options.changePasswordPath,
       profilePath = Accounts.ui._options.profilePath,
     } = this.props;
-    const { formState, waiting, user } = this.state;
+    const { user } = this.props;
+    const { formState, waiting } = this.state;
     let loginButtons = [];
 
     if (user && formState == STATES.PROFILE) {
       loginButtons.push({
         id: 'signOut',
-        label: T9n.get('signOut'),
+        label: this.translate('signOut'),
         disabled: waiting,
         onClick: this.signOut.bind(this)
       });
@@ -324,7 +329,7 @@ export class LoginForm extends Tracker.Component {
     if (this.showCreateAccountLink()) {
       loginButtons.push({
         id: 'switchToSignUp',
-        label: T9n.get('signUp'),
+        label: this.translate('signUp'),
         type: 'link',
         href: signUpPath,
         onClick: this.switchToSignUp.bind(this)
@@ -334,7 +339,7 @@ export class LoginForm extends Tracker.Component {
     if (formState == STATES.SIGN_UP || formState == STATES.PASSWORD_RESET) {
       loginButtons.push({
         id: 'switchToSignIn',
-        label: T9n.get('signIn'),
+        label: this.translate('signIn'),
         type: 'link',
         href: loginPath,
         onClick: this.switchToSignIn.bind(this)
@@ -344,22 +349,22 @@ export class LoginForm extends Tracker.Component {
     if (this.showForgotPasswordLink()) {
       loginButtons.push({
         id: 'switchToPasswordReset',
-        label: T9n.get('forgotPassword'),
+        label: this.translate('forgotPassword'),
         type: 'link',
         href: resetPasswordPath,
         onClick: this.switchToPasswordReset.bind(this)
       });
     }
 
-    if (user && !_.contains([
+    if (user && ![
         "EMAIL_ONLY_NO_PASSWORD",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields())
+      ].includes(passwordSignupFields())
       && formState == STATES.PROFILE
       && (user.services && 'password' in user.services)) {
       loginButtons.push({
         id: 'switchToChangePassword',
-        label: T9n.get('changePassword'),
+        label: this.translate('changePassword'),
         type: 'link',
         href: changePasswordPath,
         onClick: this.switchToChangePassword.bind(this)
@@ -369,7 +374,7 @@ export class LoginForm extends Tracker.Component {
     if (formState == STATES.SIGN_UP) {
       loginButtons.push({
         id: 'signUp',
-        label: T9n.get('signUp'),
+        label: this.translate('signUp'),
         type: hasPasswordService() ? 'submit' : 'link',
         className: 'active',
         disabled: waiting,
@@ -380,7 +385,7 @@ export class LoginForm extends Tracker.Component {
     if (this.showSignInLink()) {
       loginButtons.push({
         id: 'signIn',
-        label: T9n.get('signIn'),
+        label: this.translate('signIn'),
         type: hasPasswordService() ? 'submit' : 'link',
         className: 'active',
         disabled: waiting,
@@ -391,7 +396,7 @@ export class LoginForm extends Tracker.Component {
     if (formState == STATES.PASSWORD_RESET) {
       loginButtons.push({
         id: 'emailResetLink',
-        label: T9n.get('resetYourPassword'),
+        label: this.translate('resetYourPassword'),
         type: 'submit',
         disabled: waiting,
         onClick: this.passwordReset.bind(this)
@@ -401,7 +406,7 @@ export class LoginForm extends Tracker.Component {
     if (this.showPasswordChangeForm() || this.showEnrollAccountForm()) {
       loginButtons.push({
         id: 'changePassword',
-        label: (this.showPasswordChangeForm() ? T9n.get('changePassword') : T9n.get('setPassword')),
+        label: (this.showPasswordChangeForm() ? this.translate('changePassword') : this.translate('setPassword')),
         type: 'submit',
         disabled: waiting,
         onClick: this.passwordChange.bind(this)
@@ -410,7 +415,7 @@ export class LoginForm extends Tracker.Component {
       if (Accounts.user()) {
         loginButtons.push({
           id: 'switchToSignOut',
-          label: T9n.get('cancel'),
+          label: this.translate('cancel'),
           type: 'link',
           href: profilePath,
           onClick: this.switchToSignOut.bind(this)
@@ -418,7 +423,7 @@ export class LoginForm extends Tracker.Component {
       } else {
         loginButtons.push({
           id: 'cancelResetPassword',
-          label: T9n.get('cancel'),
+          label: this.translate('cancel'),
           type: 'link',
           onClick: this.cancelResetPassword.bind(this),
         });
@@ -435,7 +440,7 @@ export class LoginForm extends Tracker.Component {
           b.type != undefined);
     });
 
-    return _.indexBy(loginButtons, 'id');
+    return indexBy(loginButtons, 'id');
   }
 
   showSignInLink(){
@@ -457,11 +462,9 @@ export class LoginForm extends Tracker.Component {
   }
 
   showForgotPasswordLink() {
-    return !this.state.user
+    return !this.props.user
       && this.state.formState == STATES.SIGN_IN
-      && _.contains(
-        ["USERNAME_AND_EMAIL", "USERNAME_AND_OPTIONAL_EMAIL", "EMAIL_ONLY"],
-        passwordSignupFields());
+      && ["USERNAME_AND_EMAIL", "USERNAME_AND_OPTIONAL_EMAIL", "EMAIL_ONLY"].includes(passwordSignupFields());
   }
 
   /**
@@ -587,7 +590,7 @@ export class LoginForm extends Tracker.Component {
         error = true;
       }
       else {
-        if (_.contains([ "USERNAME_AND_EMAIL_NO_PASSWORD" ], passwordSignupFields())) {
+        if (["USERNAME_AND_EMAIL_NO_PASSWORD"].includes(passwordSignupFields())) {
           this.loginWithoutPassword();
           return;
         } else {
@@ -610,7 +613,7 @@ export class LoginForm extends Tracker.Component {
         error = true;
       }
       else {
-        if (_.contains([ "EMAIL_ONLY_NO_PASSWORD" ], passwordSignupFields())) {
+        if (["EMAIL_ONLY_NO_PASSWORD"].includes(passwordSignupFields())) {
           this.loginWithoutPassword();
           error = true;
         } else {
@@ -618,7 +621,7 @@ export class LoginForm extends Tracker.Component {
         }
       }
     }
-    if (!_.contains([ "EMAIL_ONLY_NO_PASSWORD" ], passwordSignupFields())
+    if (!["EMAIL_ONLY_NO_PASSWORD"].includes(passwordSignupFields())
       && !this.validateField('password', password)) {
       error = true;
     }
@@ -627,7 +630,7 @@ export class LoginForm extends Tracker.Component {
       Meteor.loginWithPassword(loginSelector, password, (error, result) => {
         onSubmitHook(error,formState);
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
         }
         else {
           loginResultCallback(() => this.state.onSignedInHook());
@@ -658,11 +661,12 @@ export class LoginForm extends Tracker.Component {
         });
       }
     }
-    return _.indexBy(oauthButtons, 'id');
+    return indexBy(oauthButtons, 'id');
   }
 
   oauthSignIn(serviceName) {
-    const { formState, waiting, user, onSubmitHook } = this.state;
+    const { user } = this.props;
+    const { formState, waiting, onSubmitHook } = this.state;
     //Thanks Josh Owens for this one.
     function capitalService() {
       return serviceName.charAt(0).toUpperCase() + serviceName.slice(1);
@@ -683,10 +687,11 @@ export class LoginForm extends Tracker.Component {
       options.forceApprovalPrompt = Accounts.ui._options.forceApprovalPrompt[serviceName];
 
     this.clearMessages();
+    const self = this
     loginWithService(options, (error) => {
       onSubmitHook(error,formState);
       if (error) {
-        this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"));
+        this.showMessage(`error.accounts.${error.reason}` || "unknown_error");
       } else {
         this.setState({ formState: STATES.PROFILE });
         this.clearDefaultFieldValues();
@@ -720,10 +725,10 @@ export class LoginForm extends Tracker.Component {
         options.username = username;
       }
     } else {
-      if (_.contains([
+      if ([
         "USERNAME_AND_EMAIL",
         "USERNAME_AND_EMAIL_NO_PASSWORD"
-      ], passwordSignupFields()) && !this.validateField('username', username) ) {
+      ].includes(passwordSignupFields()) && !this.validateField('username', username) ) {
         if (this.state.formState == STATES.SIGN_UP) {
           this.state.onSubmitHook("error.accounts.usernameRequired", this.state.formState);
         }
@@ -737,10 +742,10 @@ export class LoginForm extends Tracker.Component {
       options.email = email;
     }
 
-    if (_.contains([
+    if ([
       "EMAIL_ONLY_NO_PASSWORD",
       "USERNAME_AND_EMAIL_NO_PASSWORD"
-    ], passwordSignupFields())) {
+    ].includes(passwordSignupFields())) {
       // Generate a random password.
       options.password = Meteor.uuid();
     } else if (!this.validateField('password', password)) {
@@ -753,12 +758,12 @@ export class LoginForm extends Tracker.Component {
     const SignUp = function(_options) {
       Accounts.createUser(_options, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
-          if (T9n.get(`error.accounts.${error.reason}`)) {
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
+          if (this.translate(`error.accounts.${error.reason}`)) {
             onSubmitHook(`error.accounts.${error.reason}`, formState);
           }
           else {
-            onSubmitHook("Unknown error", formState);
+            onSubmitHook("unknown_error", formState);
           }
         }
         else {
@@ -804,10 +809,10 @@ export class LoginForm extends Tracker.Component {
 
       Accounts.loginWithoutPassword({ email: email }, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
         }
         else {
-          this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.showMessage(this.translate("info.emailSent"), 'success', 5000);
           this.clearDefaultFieldValues();
         }
         onSubmitHook(error, formState);
@@ -818,10 +823,10 @@ export class LoginForm extends Tracker.Component {
 
       Accounts.loginWithoutPassword({ email: usernameOrEmail, username: usernameOrEmail }, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
         }
         else {
-          this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.showMessage(this.translate("info.emailSent"), 'success', 5000);
           this.clearDefaultFieldValues();
         }
         onSubmitHook(error, formState);
@@ -829,11 +834,11 @@ export class LoginForm extends Tracker.Component {
       });
     } else {
       let errMsg = null;
-      if (_.contains([ "USERNAME_AND_EMAIL_NO_PASSWORD" ], passwordSignupFields())) {
-        errMsg = T9n.get("error.accounts.Invalid email or username");
+      if (["USERNAME_AND_EMAIL_NO_PASSWORD"].includes(passwordSignupFields())) {
+        errMsg = this.translate("error.accounts.invalid_email");
       }
       else {
-        errMsg = T9n.get("error.accounts.Invalid email");
+        errMsg = this.translate("error.accounts.invalid_email");
       }
       this.showMessage(errMsg,'warning');
       onSubmitHook(errMsg, formState);
@@ -858,10 +863,10 @@ export class LoginForm extends Tracker.Component {
 
       Accounts.forgotPassword({ email: email }, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
         }
         else {
-          this.showMessage(T9n.get("info.emailSent"), 'success', 5000);
+          this.showMessage(this.translate("info.emailSent"), 'success', 5000);
           this.clearDefaultFieldValues();
         }
         onSubmitHook(error, formState);
@@ -891,11 +896,11 @@ export class LoginForm extends Tracker.Component {
     if (token) {
       Accounts.resetPassword(token, newPassword, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
           onSubmitHook(error, formState);
         }
         else {
-          this.showMessage(T9n.get('info.passwordChanged'), 'success', 5000);
+          this.showMessage(this.translate('info.passwordChanged'), 'success', 5000);
           onSubmitHook(null, formState);
           this.setState({ formState: STATES.PROFILE });
           Accounts._loginButtonsSession.set('resetPasswordToken', null);
@@ -907,11 +912,11 @@ export class LoginForm extends Tracker.Component {
     else {
       Accounts.changePassword(password, newPassword, (error) => {
         if (error) {
-          this.showMessage(T9n.get(`error.accounts.${error.reason}`) || T9n.get("Unknown error"), 'error');
+          this.showMessage(`error.accounts.${error.reason}` || "unknown_error", 'error');
           onSubmitHook(error, formState);
         }
         else {
-          this.showMessage(T9n.get('info.passwordChanged'), 'success', 5000);
+          this.showMessage('info.passwordChanged', 'success', 5000);
           onSubmitHook(null, formState);
           this.setState({ formState: STATES.PROFILE });
           this.clearDefaultFieldValues();
@@ -921,7 +926,7 @@ export class LoginForm extends Tracker.Component {
   }
 
   showMessage(message, type, clearTimeout, field){
-    message = message.trim();
+    message = this.translate(message).trim();
     if (message) {
       this.setState(({ messages = [] }) => {
         messages.push({
@@ -994,9 +999,34 @@ export class LoginForm extends Tracker.Component {
         buttons={this.buttons()}
         {...this.state}
         message={message}
+        translate={text => this.translate(text)}
       />
     );
   }
 }
+LoginForm.propTypes = {
+  formState: PropTypes.symbol,
+  loginPath: PropTypes.string,
+  signUpPath: PropTypes.string,
+  resetPasswordPath: PropTypes.string,
+  profilePath: PropTypes.string,
+  changePasswordPath: PropTypes.string,
+};
+LoginForm.defaultProps = {
+  formState: null,
+  loginPath: null,
+  signUpPath: null,
+  resetPasswordPath: null,
+  profilePath: null,
+  changePasswordPath: null,
+};
 
 Accounts.ui.LoginForm = LoginForm;
+
+export default createContainer(() => {
+  // Listen for the user to login/logout and the services list to the user.
+  Meteor.subscribe('servicesList');
+  return ({
+    user: Accounts.user(),
+  });
+}, LoginForm);