+++ imports/ui/components/PasswordOrService.jsx
@@ -5,10 +5,6 @@ import { T9n } from 'meteor/softwarerero:accounts-t9n';
 import { hasPasswordService } from '../../helpers.js';
 
 export class PasswordOrService extends React.Component {
-  propTypes: {
-    oauthServices: PropTypes.object
-  }
-
   constructor(props) {
     super(props);
     this.state = {
@@ -17,6 +13,13 @@ export class PasswordOrService extends React.Component {
         return props.oauthServices[service].label
       })
     };
+  }
+
+  translate(text) {
+    if (this.props.translate) {
+      return this.props.translate(text);
+    }
+    return T9n.get(text);
   }
 
   render () {
@@ -30,12 +33,16 @@ export class PasswordOrService extends React.Component {
     if (hasPasswordService && services.length > 0) {
       return (
         <div style={ style } className={ className }>
-          { `${T9n.get('orUse')} ${ labels.join(' / ') }` }
+          { `${this.translate('orUse')} ${ labels.join(' / ') }` }
         </div>
       );
     }
     return null;
   }
 }
+
+PasswordOrService.propTypes = {
+  oauthServices: PropTypes.object
+};
 
 Accounts.ui.PasswordOrService = PasswordOrService;