+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'std:accounts-ui',
-  version: '1.2.20',
+  version: '1.2.22',
   summary: 'Accounts UI for React in Meteor 1.3+',
   git: 'https://github.com/studiointeract/accounts-ui',
   documentation: 'README.md'
@@ -9,13 +9,12 @@ Package.describe({
 Package.onUse(function(api) {
   api.versionsFrom('1.3');
   api.use('ecmascript');
-  api.use('tracker');
-  api.use('underscore');
   api.use('accounts-base');
   api.use('check');
   api.use('random');
   api.use('email');
   api.use('session');
+  api.use('react-meteor-data@0.2.11');
   api.use('softwarerero:accounts-t9n');
   api.use('tmeasday:check-npm-versions@0.3.0');